requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
Flask
Flask-WTF
WTForms
PyYAML
joblib
scapy
pyshark
tqdm
pyarrow
fastparquet
tensorflow
xgboost
plotly
dash